+++ source/Game/SectorHomeBase.cpp
@@ -8,8 +8,8 @@ SectorHomeBase::SectorHomeBase(SectorManager* sectormanager, const io::path & sk
 void SectorHomeBase::init(){
 	//TODO causes memory leak should create entity's
 		BasePlaceholder* base = new BasePlaceholder(this->parent);
-		base->position = irr::core::vector3df((rand()%500)-250,(rand()%250)-125,(rand()%500)-250);
-		base->orientation = irr::core::vector3df(rand()%90,rand()%90,rand()%90);
+		base->position = irr::core::vector3df((float)((rand() % 500) - 250), (float)((rand() % 250) - 125), (float)((rand() % 500) - 250));
+		base->orientation = irr::core::vector3df((float)(rand() % 90), (float)(rand() % 90), (float)(rand() % 90));
 		addComponent(base);
 	
 }
@@ -28,7 +28,10 @@ SectorHomeBase::~SectorHomeBase() {
 }
 
 BasePlaceholder::BasePlaceholder(Composite* parent):Entity(parent) {
-	this->node = Game::getSceneManager()->addCubeSceneNode(100,0,-1,irr::core::vector3df((rand()%500)-250,(rand()%250)-125,(rand()%500)-250),irr::core::vector3df(rand()%90,rand()%90,rand()%90));
+	this->node = Game::getSceneManager()->addCubeSceneNode(100, 0, -1,
+		irr::core::vector3df((float)((rand() % 500) - 250), (float)((rand() % 250) - 125), (float)((rand() % 500) - 250)),
+		irr::core::vector3df((float)(rand() % 90), (float)(rand() % 90), (float)(rand() % 90))
+	);
 		
 	this->node->setMaterialFlag(EMF_LIGHTING, false);
 	this->node->setMaterialFlag(EMF_FOG_ENABLE, true);
